Display instruction and exit installer when shell access is limited (#1668)
With no basic access to Linux commands the installer cannot complete
installation. This affects a scenario with direct UI installation
through ssh-ing into VCSA. In this scenario the installation process
involves 1) registering plugin with VC, 2) copying the plugin contents
to the Web Client cache folder, and 3) adjusting ownership of the
uploaded files. The installer script will automatically detect whether
the target VCSA has bash shell ready for the installer script and print
out detailed instructions if it is not accessible. See below for the
instructions:

WARNING!! Bash shell is required to complete the installation
automatically, however it is not enabled on this machine.
Installation cannot be completed automatically. Please follow the
steps below, or refer to instructions in the UI installation guide.

1) Make the {vic_extracted_folder}/ui/vsphere-client-serenity folder
   accessible by SSH on the system where you are running this script.

2) Connect to the target VCSA using the root account

3) When greeted with the "Command>" prompt, type the following in order:
   shell.set --enabled True
   shell

4) "scp" the folder mentioned in 1) to /etc/vmware/vsphere-client/vc-packages/
   on the target VCSA's location. Create any missing folder missing if any
   (e.g. vc-packages, vsphere-client-serenity)

5) Change owner of the vc-packages folder by typing the following command:
   chown -R vsphere-client:users /etc/vmware/vsphere-client/vc-packages

6) When all done, log out of the shell and log into the Web Client